+++ proj4rs/CHANGELOG.md
@@ -7,7 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## Unreleased
 
+### Fixed
+
+* Fix `+nadgrids=@null` as no-op on datum transformation
+
 ### Changed 
+
 * Allow 3d inputs in examples/proj4rs
 
 ### Added